HEC-480: Extract Runtime into discrete modules
ExtensionDispatch (extend, capability, apply_hecksagon_capabilities),
ConfigurationDSL (use, adapter, enable, swap_adapter),
CommandDispatch (run, dry_run, on, events, []).
Runtime.rb: 330 → 104 lines.
